[server] change type HoursSpent

+++ server/Timelogger/Entities/Activity.cs
@@ -9,7 +9,7 @@ namespace Timelogger.Entities {
         public int ProjectId { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public decimal HoursSpent { get; set; }
+        public int HoursSpent { get; set; }
 
     }
 }
